chore: size smaller
